+++ pos_next/api/offers.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2025, POS Next and contributors
+# Copyright (c) 2025, FatehPOS and contributors
 # For license information, please see license.txt
 
 """